Part4 to 4.22*

+++ part4/controllers/blogs.js
@@ -1,20 +1,26 @@
 const Blog = require('../models/Blog')
 const User = require('../models/User')
 const blogsRouter = require('express').Router()
+const middlewares = require('../utils/middlewares')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
   
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middlewares.checkAndGetUserToken, async (request, response) => {
   const newBlogData = {...request.body}
 
   if(!newBlogData.title || !newBlogData.url)
     return response.status(400).json({ error: 'title or url missing' })
 
-  const users = await User.find({})
-  const userAuthor = users[0]
+  const userAuthor = await User.findById(newBlogData.userLogin)
+
+  if(!userAuthor) {
+    return response.status(404).json({ error: 'User not found' })
+  }
+
+  delete newBlogData.userLogin
 
   newBlogData.user = userAuthor.id
 
@@ -47,7 +53,13 @@ blogsRouter.put('/:id', async (request, response) => {
   response.status(200).json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middlewares.checkAndGetUserToken, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  
+  if(blog.user.toString() !== request.body.userLogin) { 
+    return response.status(401).json({ error: 'unauthorized' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
